+++ backend/app/schemas/limit.py
@@ -1,10 +1,12 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class QuarterlyLimitResponse(BaseModel):
     year: int
     quarter: int
-    limit: float
-    used: float
-    remaining: float
+    limit: Decimal
+    used: Decimal
+    remaining: Decimal
     is_exceeded: bool
